Return to a single js file structure
