+++ index.html
@@ -12,7 +12,7 @@
     <script src="https://www.gstatic.com/firebasejs/9.6.0/firebase-database-compat.js"></script>
 
     <!-- Load the Game main file  -->
-    <script type="module" src="assets/js/main.js"></script>
+    <script type="module" src="assets/js/scripts.js"></script>
 </head>
 
 <body>
